Trata possíveis erros de usuário e exceções

Approved-by: Manoel Vilela <manoel_vilela@engineer.com>

+++ teca/admin.py
@@ -9,23 +9,27 @@ from mysql.connector.errors import DatabaseError
 
 def admin_ler_entrada(atributo):
     if atributo == 'senha_hash':
-        valor = getpass.getpass(prompt="senha: ")
+        valor = database.senha_hash(getpass.getpass(prompt="senha: "))
     elif 'data' in atributo:
         valor = check.entrada('>>> ', check.data)
     elif atributo == 'nickname':
         valor = check.entrada('>>> ', check.nickname)
     elif atributo == 'numero':
         valor = check.entrada('>>> ', check.telefone)
+    elif 'cpf' in atributo:
+        valor = check.entrada('>>> ', check.cpf)
+    elif 'isbn' in atributo:
+        valor = check.entrada('>>> ', check.isnb)
     else:
         valor = check.entrada('>>> ', check.nao_vazia)
 
     return valor
 
 
-def escolher_tabela():
+def escolher_tabela(t):
     print("Escolha uma das tabelas: ")
     tabelas = {str(idx+1): tabela
-               for idx, tabela in enumerate(database.tabelas)}
+               for idx, tabela in enumerate(t)}
     menu = {k: v._table for k, v in tabelas.items()}
     escolha = term.menu_enumeracao(menu)
     return tabelas[escolha]
@@ -46,25 +50,54 @@ def escolher_tupla(tabela):
 
 def admin_inserir():
     print("== INSERIR")
-    tabela_escolhida = escolher_tabela()
+    tabela_escolhida = escolher_tabela(database.tabelas_sem_isa)
     atributos = []
     print("Inserção dos atributos na tabela: ", tabela_escolhida._table)
     for nome_atributo in tabela_escolhida._columns:
-        print(f'{nome_atributo}: ')
+        if nome_atributo != 'senha_hash':
+            print(f'{nome_atributo}: ')
         entrada = admin_ler_entrada(nome_atributo)
         atributos.append(entrada)
 
     instancia = tabela_escolhida(*atributos)
+
+    if tabela_escolhida._table == 'usuario':
+        if instancia.tipo == 'aluno':
+            tabela = database.Aluno
+        elif instancia.tipo == 'funcionario':
+            tabela = database.Funcionario
+        elif instancia.tipo == 'professor':
+            tabela = database.Professor
+        else:
+            print("Tipo de usuário inválido!")
+            return
+        atributos = [instancia.matricula]
+        print("Inserção dos atributos na tabela: ", tabela._table)
+        for nome_atributo in tabela._columns[1:]:
+            print(f'{nome_atributo}: ')
+            entrada = admin_ler_entrada(nome_atributo)
+            atributos.append(entrada)
+
+        usuario_extra = tabela(*atributos)
+
     try:
-        instancia.insert()
+        status = instancia.insert()
+        if tabela_escolhida._table == 'usuario' and status:
+            usuario_extra.insert()
     except DatabaseError as e:
         print("Não foi possível completar a ação. Uma exceção foi disparada!")
         print("Exceção: ", e)
+        if status:
+            instancia.delete()
+        return
+
+    if status:
+        print("INSERÇÃO FINALIZADA COM SUCESSO!")
 
 
 def admin_alterar():
     print("== ALTERAR")
-    tabela_escolhida = escolher_tabela()
+    tabela_escolhida = escolher_tabela(database.tabelas_todas)
     instancia = escolher_tupla(tabela_escolhida)
     atributos = {str(idx+1): attr
                  for idx, attr in enumerate(tabela_escolhida._columns)}
@@ -79,21 +112,26 @@ def admin_alterar():
 
 def admin_remover():
     print("== REMOVER")
-    tabela_escolhida = escolher_tabela()
+    tabela_escolhida = escolher_tabela(database.tabelas_sem_isa)
     instancia = escolher_tupla(tabela_escolhida)
-    instancia.delete()
+    if instancia is not None:
+        deleted = instancia.delete()
+        if deleted:
+            print("TUPLA DELETADA COM SUCESSO!")
+    else:
+        print("TUPLA NÃO ENCONTRADA.")
 
 
 def admin_imprimir():
     print("== IMPRIMIR")
-    tabela_escolhida = escolher_tabela()
+    tabela_escolhida = escolher_tabela(database.tabelas_todas)
     term.imprimir_tabela(tabela_escolhida)
 
 
 def tela_admin():
+    print("== TELA DE ADMINISTRADOR ==")
     while True:
         opcoes = {
-            '0': 'Sair',
             '1': 'Inserir',
             '2': 'Remover',
             '3': 'Alterar',
@@ -101,13 +139,16 @@ def tela_admin():
         }
         print("Opções: ")
         opcao = term.menu_enumeracao(opcoes)
-        if opcao == '0':
-            break
-        elif opcao == '1':
-            admin_inserir()
-        elif opcao == '2':
-            admin_remover()
-        elif opcao == '3':
-            admin_alterar()
-        elif opcao == '4':
-            admin_imprimir()
+        try:
+            if opcao == '0':
+                break
+            elif opcao == '1':
+                admin_inserir()
+            elif opcao == '2':
+                admin_remover()
+            elif opcao == '3':
+                admin_alterar()
+            elif opcao == '4':
+                admin_imprimir()
+        except KeyboardInterrupt:
+            print("\nOperação interrompida!")
